Add REST interface to access the current month balance of a user

+++ src/main/java/com/financeplanner/config/security/jwt/TokenAuthenticationFilter.java
@@ -35,7 +35,7 @@ public class TokenAuthenticationFilter extends OncePerRequestFilter {
             String jwt = getJwtFromRequest(request);
 
             if (StringUtils.hasText(jwt) && tokenProvider.validateToken(jwt)) {
-                Long userId = tokenProvider.getUserIdFromToken(jwt);
+                int userId = tokenProvider.getUserIdFromToken(jwt).intValue();
 
                 UserDetails userDetails = customUserDetailsService.loadUserById(userId);
                 UsernamePasswordAuthenticationToken authentication
